chore: re-organise match module
In particular, splitting up the class, functions and types into separate
modules.
